finished lab

+++ lab_hash/dhhashtable.cpp
@@ -84,7 +84,7 @@ void DHHashTable<K, V>::insert(K const& key, V const& value)
     (void) key;   // prevent warnings... When you implement this function, remove this line.
     (void) value; // prevent warnings... When you implement this function, remove this line.
     elems++;
-    if (elems/size >= 0.7)
+    if ((double)elems/(double)size >= 0.7)
     resizeTable();
     // size_t i = hashes::hash(key,size);
     // size_t j = hashes::secondary_hash(key,size);
